Updated build script to retrieve and build current stable release version (v7.6.0) as of 07-Dec-2017...
All dependencies that have hard-coded branches/releases have been
checked against the Maven "pom.xml" files referenced in the comments,
and updated accordingly.

The glitches/TO-DO's from the v7.4.1 version of this repo have not yet
been resolved, so several manual actions still need to occur to remove
the SAP dependency after the script has finished running.

As it's very easy to press a key at the end of the Maven build and have
these instructions scroll off the screen, they have been copied over to
the README.md file.